\name{as.zoo.timets}
\alias{as.zoo.timets}
\title{convert form a TIME time series to a zoo series}
\usage{
  as.zoo.timets(x, ...)
}
\arguments{
  \item{x}{an object of class 'timets'}

  \item{...}{arguments passed to the zoo constructor}
}
\value{
  a zoo DAILY time series
}
\description{
  convert form a TIME time series to a zoo series
}
\seealso{
  /code{/link{timets}}
}

