\name{createModelRunner}
\alias{createModelRunner}
\title{Create a TIME Model runner}
\usage{
  createModelRunner(type)
}
\arguments{
  \item{type}{a System.Type object, type of the model in
  TIME}
}
\value{
  a clr object implementing IPointTimeSeriesSimulation
}
\description{
  Create a TIME Model runner
}

