\name{createTts}
\alias{createTts}
\title{Create a 'TIME' time series}
\usage{
  createTts(values, startDate, timestep)
}
\arguments{
  \item{values}{the numeric vector of values in the time
  series}

  \item{startDate}{an R Date object, start date}

  \item{timeStep}{a case-sensitive string such as "Daily",
  "Monthly"}
}
\value{
  a 'TIME' time series
}
\description{
  Create a 'TIME' time series
}

