\name{.onLoad}
\alias{.onLoad}
\title{Function called when loading the RtoTIME package with 'library'.}
\usage{
  .onLoad(libname = "~/R", pkgname = "RtoTIME")
}
\arguments{
  \item{libname}{the path to the library from which the
  package is loaded}

  \item{pkgname}{the name of the package.}
}
\description{
  Function called when loading the RtoTIME package with
  'library'.
}

