\name{getModelParameters}
\alias{getModelParameters}
\title{Gets a TIME parameter set from a model}
\usage{
  getModelParameters(model)
}
\arguments{
  \item{model}{A TIME model, usually; possibly an arbitrary
  CLR object}
}
\value{
  a CLR object of type ParameterSet
}
\description{
  Gets a TIME parameter set from a model
}

