\name{getModelParametersAsDataFrame}
\alias{getModelParametersAsDataFrame}
\title{Gets a simple data frame representation of model parameter}
\usage{
  getModelParametersAsDataFrame(model)
}
\arguments{
  \item{model}{A TIME model, usually; possibly an arbitrary
  CLR object}
}
\value{
  a data frame with column names Name,Value,Min,Max
}
\description{
  Gets a simple data frame representation of model
  parameter
}

