\name{loadParameterSetFromXml}
\alias{loadParameterSetFromXml}
\title{Load a parameter set from an XML file}
\usage{
  loadParameterSetFromXml(filename)
}
\arguments{
  \item{filename}{the file name of the XML werialisation of
  a TIME parameter set}
}
\value{
  a CLR object
}
\description{
  Load a parameter set from an XML file
}

