\name{loadSystemRunner}
\alias{loadSystemRunner}
\title{Load a system runner simulation from an XML file}
\usage{
  loadSystemRunner(filename)
}
\arguments{
  \item{filename}{path to XML serialization of a TIME
  simulation}
}
\value{
  a clr object implementing IPointTimeSeriesSimulation
}
\description{
  Load a system runner simulation from an XML file
}

