\name{pSetApplyDataFrame}
\alias{pSetApplyDataFrame}
\title{Gets a simple data frame representation of model parameter}
\usage{
  pSetApplyDataFrame(model, paramSetDataFrame)
}
\arguments{
  \item{model}{A TIME model, usually; possibly an arbitrary
  CLR object}

  \item{paramSetDataFrame}{a data frame with at least
  column names Name,Value}
}
\description{
  Gets a simple data frame representation of model
  parameter
}

