\name{pSetAsDataFrame}
\alias{pSetAsDataFrame}
\title{Convert a TIME parameter to a simple data frame representation}
\usage{
  pSetAsDataFrame(paramSet)
}
\arguments{
  \item{paramSet}{a CLR object of type ParameterSet}
}
\value{
  a data frame with column names Name,Value,Min,Max
}
\description{
  Convert a TIME parameter to a simple data frame
  representation
}

