\name{rewindModelRunner}
\alias{rewindModelRunner}
\title{Rewinds, reset and initialise states of a ModelRunner.}
\usage{
  rewindModelRunner(simul)
}
\arguments{
  \item{simul}{CLr object of type ModelRunner}
}
\description{
  Rewinds the model, then call reset and
  initialiseSimulation on the model that the ModelRunner
  uses.
}

