\docType{package}
\name{RtoTIME-package}
\alias{RtoTIME}
\alias{RtoTIME-package}
\title{Access TIME high-level modelling constructs from R}
\description{
  Access TIME high-level modelling constructs from R
}
\details{
  \tabular{ll}{ Package: \tab RtoTIME\cr Type: \tab
  Package\cr Version: \tab 0.2-1\cr Date: \tab
  2013-09-22\cr License: \tab LGPL 2.1\cr }
}
\author{
  Jean-Michel Perraud
  \email{jean-michel.perraud_at_csiro.au}
}
\keyword{.NET}
\keyword{CLR}
\keyword{Mono}
\keyword{TIME}
\keyword{package}

