\name{runTimeStep}
\alias{runTimeStep}
\title{Run one time step of a ModelRunner temporal simulation}
\usage{
  runTimeStep(modelRunner, sDate)
}
\arguments{
  \item{modelRunner}{a TIME model runner}

  \item{sDate}{the simulation date}
}
\description{
  Run one time step of a ModelRunner temporal simulation
}

