\name{setSimulSpan}
\alias{setSimulSpan}
\title{Sets the temporal span of the model simulation}
\usage{
  setSimulSpan(simulation, startDate, endDate)
}
\arguments{
  \item{simulation}{a TIME ModelRunner object}

  \item{startDate}{The start of the simulation}

  \item{endDate}{last day of the simulation}
}
\description{
  Sets the temporal span of the model simulation
}

