\name{timets}
\alias{timets}
\title{OBSOLETE create a simple 'TIME' time series}
\usage{
  timets(x, startdate)
}
\arguments{
  \item{x}{a numeric vector}

  \item{startdate}{arguments passed to the zoo constructor}
}
\value{
  a very simple time series structure: data and start date
}
\description{
  OBSOLETE create a simple 'TIME' time series
}

