\name{zooToDailyTts}
\alias{zooToDailyTts}
\title{Convert a 'zoo' series to a 'TIME' time series}
\usage{
  zooToDailyTts(zoots)
}
\arguments{
  \item{zoots}{a zoo time series}
}
\value{
  a clrobj object of type TIME.DataTypes.TimeSeries
}
\description{
  Convert a 'zoo' series to a 'TIME' time series
}

